add dependencies and refactor common codes into a utils package
