+++ fastmod.py
@@ -3,40 +3,21 @@ import numpy as np
 import scipy as sp
 import math
 import time
-from scipy.interpolate import RectBivariateSpline
+from utils import t6hat, sin6hat
 
-def t6hat(rmax,rmin,x):
-#======================
-   xx   = np.abs(x)
-   ones = np.ones_like(xx)
-   xmax = rmax*ones
-   xmin = rmin*ones
+# --------- define polar coordinate transformation, readgeom, readinverse and prepare_hankels routines
 
-   xx1 = np.minimum(xx,xmax)
-   xx2 = np.maximum(xx1,xmin)
-   xx3 = (ones - (xx2 - rmin)/(rmax-rmin))*math.pi
-   xx4 = sin6hat(xx3)
-   return xx4
-
-def sin6hat(x):
-#==============
-   res = np.sin(6.0*x)/3.0 - 3.0*np.sin(4.0*x)   + 15.0*np.sin(2.0*x) - 20.0*x
-   res = -res/(20.0*math.pi)
-   return res
-
-#--------- define polar coordinate transformation, readgeom, readinverse and prepare_hankels routines
-
-def cart_to_polar_coor(xx,yy):               #------------------------- begin cart_to_polar
-#=============================
+def cart_to_polar_coor(xx,yy):               # ------------------------- begin cart_to_polar
+# =============================
    rad = np.sqrt( xx*xx + yy*yy  )
    angle = np.arccos(xx/(rad+1.e-300))
    angle[yy < 0.] = 2.* math.pi - angle[yy < 0.]
 
-   return rad,angle                          #------------------------- end cart_to_polar
+   return rad, angle                          # ------------------------- end cart_to_polar
 
 
-def prepare_hankels(ndet,numfreq,freq_rad):     #---------------------- begin prepare_hankels
-#==========================================
+def prepare_hankels(ndet,numfreq,freq_rad):     # ---------------------- begin prepare_hankels
+# ==========================================
    global hankels
 
    if('hankels' in globals()):
@@ -68,7 +49,7 @@ def prepare_hankels(ndet,numfreq,freq_rad):     #---------------------- begin pr
        hankels[0,:]  =   large_number
        hankels[:,0] = large_number
 
-   #return hankels                                #--------------------- end prepare hankels
+   #return hankels                                # --------------------- end prepare hankels
 
 
 def find_good_ntimes(ngood):
@@ -79,10 +60,9 @@ def find_good_ntimes(ngood):
 
    index = 1 + np.array(range(18))
    goodnumbers = np.insert(num2pow, index, num2pow3)
-   #print(goodnumbers)
 
-   for j  in goodnumbers:
-       if j>ngood:
+   for j in goodnumbers:
+       if j > ngood:
           break
 
    newtimes = j+1
@@ -90,7 +70,7 @@ def find_good_ntimes(ngood):
    return newtimes
 
 def fast_inverse_mod(data,ndet,ntimes,detector_radius,t_measured,speed_of_sound,ndim,lhalf,rad_corr):      # ================= T H I S    I S    T H E    M A I N     F U N C T I O N  ======================
-#====================================================================================================        --------------------------------------------------------------------------------------------------
+# ====================================================================================================        --------------------------------------------------------------------------------------------------
     global hankels
 
     start_time = time.time()
@@ -101,21 +81,20 @@ def fast_inverse_mod(data,ndet,ntimes,detector_radius,t_measured,speed_of_sound,
 
 
     tmax = t_measured/detector_radius*speed_of_sound          #  this is Tmax scaled to a problem with R = 1, speed_of_sound = 1
-    tstep = tmax/(ntimes - 1)
+    tstep = tmax / (ntimes - 1)
 
-    ngood = 3.1/tstep +1
-    #ngood = 5.1/tstep +1
+    ngood = 3.1 / tstep +1
 
     newtimes = find_good_ntimes(ngood)
 
-    tmax = (newtimes-1)*tstep
+    tmax = (newtimes-1) * tstep
 
     print('--------- newtimes = ',newtimes,'  new tmax = ',tmax)
 
     datanew = np.zeros((newtimes,ndet))
 
     if(newtimes > ntimes):
-        datanew[0:ntimes,:] = np.copy(data)
+        datanew[0:ntimes, :] = np.copy(data)
     else:
         if(newtimes < ntimes):
            datanew = np.copy(data[0:newtimes,:])
@@ -151,25 +130,25 @@ def fast_inverse_mod(data,ndet,ntimes,detector_radius,t_measured,speed_of_sound,
 
 
 
-    #----------------------- make a smooth cut of the input data ------------------------
+    # ----------------------- make a smooth cut of the input data ------------------------
     #tcut = 0.9*tmax
     #hat = t6hat(tmax/l,tcut/l,times)
     #hat = hat[0:nfft_points]
 
-    cdata = np.zeros((nfft_points,ndet), dtype='complex')
+    cdata = np.zeros((nfft_points, ndet), dtype='complex')
 
     for j in range(ndet):
        cdata[:,j] = np.copy(data[0:nfft_points,j]) #*hat
 
     print(np.shape(cdata))
-    #-------------------------- prepare Hankel functions -------------------------------------
+    # -------------------------- prepare Hankel functions -------------------------------------
 
     prepare_hankels(ndet,numfreq,freq_rad)
 
     hankel_time = time.time()
     print('Preparing Hankel functions','{:05.2f}'.format(hankel_time - start_time),' sec')
 
-    #----------------------- fft in time and in angle ----
+    # ----------------------- fft in time and in angle ----
 
     ctemp_fourier =  sp.fft.fftshift( sp.fft.ifft2(cdata), [1])
     cfourier = ctemp_fourier[0:numfreq]
@@ -177,7 +156,7 @@ def fast_inverse_mod(data,ndet,ntimes,detector_radius,t_measured,speed_of_sound,
     time_fft = time.time()
     print('FFT data in time and angle  ','{:05.2f}'.format(time_fft - hankel_time),' sec')
 
-    #--------------------------- multiplying by (-i)^k ---------------------------------------
+    # --------------------------- multiplying by (-i)^k ---------------------------------------
     midk = round(ndet/2)
 
     for kk in range(ndet):
@@ -186,18 +165,18 @@ def fast_inverse_mod(data,ndet,ntimes,detector_radius,t_measured,speed_of_sound,
        cfourier[:,kk] = cfourier[:,kk]/coef    # divide or multiply ?
 
 
-    #----------------------------- divide Fourier transform by Hankels --------------------------------
+    # ----------------------------- divide Fourier transform by Hankels --------------------------------
 
     cfourier_new = cfourier/hankels
 
-    #------------------------------------ Finding the DC number by integrating the zero harmonics -----
+    # ------------------------------------ Finding the DC number by integrating the zero harmonics -----
 
     czero_harm = np.copy(cfourier_new[:,midk])
 
     j1 =  np.real(hankels[:,midk+1])
     csum = np.sum(czero_harm[1:]*j1[1:]/freq_rad)*freqstep  # *2/math.pi
 
-    #----------------------------- inverse Fourier transform in angle ---------------------------------
+    # ----------------------------- inverse Fourier transform in angle ---------------------------------
 
     ctran =  sp.fft.fft(   sp.fft.fftshift( cfourier_new, [1]),axis = 1)
 
@@ -206,7 +185,7 @@ def fast_inverse_mod(data,ndet,ntimes,detector_radius,t_measured,speed_of_sound,
     time_fft = time.time()
     print('Inverse Fourier in angle','{:05.2f}'.format(time_fft - hankel_time ),' sec')
 
-    #------------------------ interpolate from the polar grid to Cartesian in the Fourier domain ------
+    # ------------------------ interpolate from the polar grid to Cartesian in the Fourier domain ------
 
 
     freq_scale =  np.linspace(-spatial_freq_max,spatial_freq_max,ndim)
@@ -222,15 +201,15 @@ def fast_inverse_mod(data,ndet,ntimes,detector_radius,t_measured,speed_of_sound,
 
     start_spline_time = time.time()
 
-    interp_spline_r = RectBivariateSpline(time_freq_scale,det_angle_ext,np.real(ctran_ext),kx=3,ky=3)
-    interp_spline_i = RectBivariateSpline(time_freq_scale,det_angle_ext,np.imag(ctran_ext),kx=3,ky=3)
+    interp_spline_r = sp.interpolate.RectBivariateSpline(time_freq_scale,det_angle_ext,np.real(ctran_ext),kx=3,ky=3)
+    interp_spline_i = sp.interpolate.RectBivariateSpline(time_freq_scale,det_angle_ext,np.imag(ctran_ext),kx=3,ky=3)
 
     cart_transform = interp_spline_r.ev(freq_rho,freq_teta) + 1j * interp_spline_i.ev(freq_rho,freq_teta)
 
     spline_time = time.time()
 
     print('Interpolation to Cartesian grid ','{:05.2f}'.format(spline_time - start_spline_time),' sec')
-    #------------------------- cut off redundant frequencies -------------------------------------------
+    # ------------------------- cut off redundant frequencies -------------------------------------------
     if(lhalf == 1):
        mcenter = round((ndim-1)/2)
        cart_transform[0:mcenter,:] = np.conj(np.copy(np.flip(cart_transform[mcenter+1:,:],[0,1])))
@@ -240,7 +219,7 @@ def fast_inverse_mod(data,ndet,ntimes,detector_radius,t_measured,speed_of_sound,
     0:mcenter,:],[0,1])))
 
 
-    #------------------------- inverse 2D Fourier transform to obtain the image ------------------------
+    # ------------------------- inverse 2D Fourier transform to obtain the image ------------------------
     coefficient = tmax  # ------ why?
 
     image =  coefficient* np.real( sp.fft.fftshift( sp.fft.fft2(   sp.fft.fftshift( cart_transform[0:ndim-1,0:ndim-1], [0,1])), [0,1] )   )
@@ -251,7 +230,7 @@ def fast_inverse_mod(data,ndet,ntimes,detector_radius,t_measured,speed_of_sound,
 
     print('2D inverse FFT ','{:05.2f}'.format(fft_time - spline_time),' sec')
 
-    #-------------------------- additional correction for zero boundary conditions ----------------------
+    # -------------------------- additional correction for zero boundary conditions ----------------------
 
     xx = np.linspace(-halfsize,halfsize,ndim)
     xstep = xx[1]-xx[0]
@@ -272,7 +251,7 @@ def fast_inverse_mod(data,ndet,ntimes,detector_radius,t_measured,speed_of_sound,
 
     image_ext[rad >= 1.] = 0.      # set to 0 outside of the unit circle
 
-    #-------------------------------- done --------------------------------------------------------------
+    # -------------------------------- done --------------------------------------------------------------
 
     final_time = time.time()
