+++ simulations/partial_data/simul_demo.py
@@ -1,34 +1,17 @@
-#----------------------------------------
+# ----------------------------------------
 #  This is a demo file illustrating the use of the "fast inverse" module
 #   for reconstructing an image from a real mouse scan
 #
-#-----------------------------------------------------------------------------------------------
+# -----------------------------------------------------------------------------------------------
 import sys
 sys.path.append("../../")
 import numpy as np
-from utils import ndwrite
+from utils import npyread
 from fastmod import fast_inverse_mod
 
-
-def npyread(fpath):
-   data = np.load(fpath)
-   #print(np.shape(data))
-   #print(type(data))
-   arr = np.zeros((2,2))
-
-   if(type(data) != type(arr)):
-      print(fpath+' is not a 2-dim array !')
-      print('Shutting down !')
-   return data
-
-
-
-
-
-#-----------------------------------------------------------------------------------------------------
+# -----------------------------------------------------------------------------------------------------
 #  Main program starts here
-#-----------------------------------------------------------------------------------------------------
-
+# -----------------------------------------------------------------------------------------------------
 
 print(' ')
 print('**********')
@@ -36,7 +19,7 @@ print('Simul demo')
 print('**********')
 print(' ')
 
-#---- data parameters:
+# ---- data parameters:
 
 detector_radius = 1.    # in meters
 speed_of_sound  = 1.      # meters per second
@@ -56,7 +39,7 @@ time_step =      1./sampling_rate              # time step of measurements
 t_measured      = time_step*(nsamples-1)        # this is total measurement time
 
 
-#----------------------------------------------- read the mouse data and transpose the matrix
+# ----------------------------------------------- read the mouse data and transpose the matrix
 data = npyread(namedata)
 (nrow,ncol) = np.shape(data)
 
@@ -71,7 +54,7 @@ if (ncol != ndet):
     print('Wrong number of detectors !')
     quit()
 
-#--------------------------- call the module to reconstruct the image ---------------
+# --------------------------- call the module to reconstruct the image ---------------
 
 lhalf = 0                   # additional filtration to correct for absent detectors on the left (0 = none)
 
@@ -83,6 +66,4 @@ lhalf = 2                   # additional filtration to correct for absent detect
 image = fast_inverse_mod(data,ndet,nsamples,detector_radius,t_measured,speed_of_sound,ndim,lhalf,rad_corr)
 np.save('reconstruction_with_corrections.npy',image)
 
-
 quit()
-
